- Removed a useless test that can't pass anyway.
The testcase setUp does::

    ezcSignalStaticConnections::getInstance()->connections = array();

Which initializes the ezcSignalStaticConnections::$instance.

testDelayedInitialization calls ezcBaseInit::setCallback, and tests that the
callback was called and works. Of course the callback never gets called as the
private ezcSignalStaticConnections::$instance was set in setUp().

This test is actually a test for ezcBaseInit and not SignalSlot itself. As a
matter of fact, removing this test does not affect code coverage.