- Bugfix: some static connections were left over.
If the signal collection had a $staticConnectionsHolder, it was ignoring
static connections from the singleton static connection holder.

According to the tests, the singleton static connection holder is not to
be ignored.

As a solution, emit() now honorates connections from the local
static connection holder and the singleton connection holder.